requirements.txt
requests
geopandas
shapely
flask
pillow